requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
appnope==0.1.3
    # via ipython
asgiref==3.7.2
    # via django
asttokens==2.4.0
    # via stack-data
backcall==0.2.0
    # via ipython
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.3
    # via pyjwt
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via python3-openid
django==4.2.5
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-htmx
    #   django-sql-utils
    #   django-stubs
    #   django-stubs-ext
django-allauth==0.56.1
    # via -r requirements.in
django-htmx==1.16.0
    # via -r requirements.in
django-sql-utils==0.6.1
    # via -r requirements.in
django-stubs[compatible-mypy]==4.2.4
    # via -r requirements.in
django-stubs-ext==4.2.2
    # via django-stubs
executing==1.2.0
    # via stack-data
idna==3.4
    # via requests
ipython==8.15.0
    # via -r requirements.in
jedi==0.19.0
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
mypy==1.5.1
    # via django-stubs
mypy-extensions==1.0.0
    # via mypy
oauthlib==3.2.2
    # via requests-oauthlib
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.39
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.16.1
    # via ipython
pyjwt[crypto]==2.8.0
    # via django-allauth
python-dotenv==1.0.0
    # via -r requirements.in
python3-openid==3.2.0
    # via django-allauth
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via asttokens
sqlparse==0.4.4
    # via
    #   django
    #   django-sql-utils
stack-data==0.6.2
    # via ipython
traitlets==5.10.0
    # via
    #   ipython
    #   matplotlib-inline
types-pytz==2023.3.1.1
    # via django-stubs
types-pyyaml==6.0.12.12
    # via django-stubs
typing-extensions==4.8.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
urllib3==2.0.4
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
